deployment/requirements.txt
snowflake-connector-python
snowflake-snowpark-python
scikit-learn
joblib
pandas
numpy
streamlit